Import hooks from main `react-relay` file (#347)
All hook imports are available from `react-relay/index.js` (specified in
`main` in the `package.json`) as of react-relay 11.0.0.

Bundlers have no issues in resolving `react-relay/hooks` to
`react-relay/hooks.js` but NodeJs `require` or `import` calls are more
strict, which means that `import * as Hooks from "react-relay/hooks"`
fails in a NodeJs environment.

By importing from the package's primary entry point we ensure that
Node's file resolver is also happy.

The peer dependency is bumped to require at least react-relay 11.0.0
which introduced the change. `=>` is used because it keeps the
open-ended upper bound that `*` also had.